// client/src/models/offers.ts
// Generated by https://quicktype.io

export interface Offers {
    buy:     Buy[];
    content: Content;
    rent:    Buy[];
    stream:  { [key: string]: string[] };
}

export interface Buy {
    location: string;
    price:    number;
    url:      string;
}

export interface Content {
    age_certification:     string;
    backdrops:             Backdrop[];
    clips:                 Clip[];
    credits:               Credit[];
    external_ids:          ExternalID[];
    full_path:             string;
    full_paths:            FullPaths;
    genre_ids:             number[];
    id:                    number;
    jw_entity_id:          JwEntityID;
    object_type:           string;
    offers:                Offer[];
    original_release_year: number;
    original_title:        string;
    permanent_audiences:   string[];
    poster:                string;
    poster_blur_hash:      string;
    production_countries:  string[];
    runtime:               number;
    scoring:               Scoring[];
    short_description:     string;
    title:                 string;
}

export interface Backdrop {
    backdrop_blur_hash: string;
    backdrop_url:       string;
}

export interface Clip {
    external_id: string;
    name:        string;
    provider:    string;
    type:        string;
}

export interface Credit {
    character_name?: string;
    name:            string;
    person_id:       number;
    role:            string;
}

export interface ExternalID {
    external_id: string;
    provider:    string;
}

export interface FullPaths {
    MOVIE_DETAIL_OVERVIEW: string;
}

export enum JwEntityID {
    Tm57536 = "tm57536",
}

export interface Offer {
    country:                       Country;
    currency:                      Currency;
    jw_entity_id:                  JwEntityID;
    monetization_type:             MonetizationType;
    package_short_name:            string;
    presentation_type:             PresentationType;
    provider_id:                   number;
    retail_price?:                 number;
    urls:                          Urls;
    last_change_date?:             string;
    last_change_date_provider_id?: string;
    last_change_difference?:       number;
    last_change_percent?:          number;
    last_change_retail_price?:     number;
    available_to?:                 string;
}

export enum Country {
    No = "NO",
}

export enum Currency {
    Nok = "NOK",
}

export enum MonetizationType {
    Buy = "buy",
    Flatrate = "flatrate",
    Rent = "rent",
}

export enum PresentationType {
    HD = "hd",
    SD = "sd",
}

export interface Urls {
    deeplink_rokuos?: string;
    standard_web:     string;
    deeplink_web?:    string;
}

export interface Scoring {
    provider_type: string;
    value:         number;
}